fixed PR recommendations
