+++ client/src/components/ReviewsShowPage.js
@@ -1,45 +1,12 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import ErrorList from "./registration/components/layout/ErrorList";
-import translateServerErrors from "../services/translateServerErrors"
 import NewReviewForm from "./NewReviewForm";
 
-const ReviewsShowPage = ({ id, currentUser, activityName, reviews, reviewList, setReviewList }) => {
-
-  const [errors, setErrors] = useState([])
-
-  const postReview = async (newReview) => {
-    try {
-      const response = await fetch(`/api/v1/activities/${id}/reviews`, {
-        method: "POST",
-        headers: new Headers({
-          "Content-Type": "application/json"
-        }),
-        body: JSON.stringify(newReview)
-      })
-      if (!response.ok) {
-        if (response.status === 422) {
-          const errorBody = await response.json()
-          const newErrors = translateServerErrors(errorBody.errors)
-          return setErrors(newErrors)
-        } else {
-          const errorMessage = `${response.status} (${response.statusText})`
-          const error = new Error(errorMessage)
-          throw error
-        }
-      } else {
-        const reviewData = await response.json()
-        setErrors([])
-        setReviewList([...reviewList, reviewData.review])
-      }
-    } catch (error) {
-      console.error(`Error in fetch: ${error.message}`)
-    }
-  }
+const ReviewsShowPage = ({ activityName, reviews, postReview, errors }) => {
 
   const listOfReviews = reviews.map(({ id, content }) => {
     return <li key={id}>{content}</li>;
   });
-
 
   let reviewContent
   if (listOfReviews.length === 0) {
@@ -54,7 +21,7 @@ const ReviewsShowPage = ({ id, currentUser, activityName, reviews, reviewList, s
         <h5>Share your thoughts on {activityName} !!</h5>
         <NewReviewForm
           postNewReview={postReview}
-          currentUser={currentUser} />
+        />
       </div>
       <div className="reviews">
         <h4>Reviews</h4>
