+++ server/src/routes/api/v1/activityRouter.js
@@ -1,14 +1,15 @@
 import express from "express"
 import { Activity } from "../../../models/index.js"
 import ActivitySerializer from "../../../serializers/ActivitySerializer.js"
-import reviewRouter from "./reviewRouter.js"
+import activityReviewRouter from "./activityReviewRouter.js"
 
 const activityRouter = new express.Router()
 
-activityRouter.use("/:id/reviews", reviewRouter)
+activityRouter.use("/:id/reviews", activityReviewRouter)
 
 activityRouter.get("/:id", async (req, res) => {
     const activityId = req.params.id
+
     try {
         const activity = await Activity.query().findById(activityId)
         const serializedActivity = await ActivitySerializer.getSummaryOfOne(activity)
